move `window/main` package to outside
