+++ susapad/window/__init__.py
@@ -7,7 +7,7 @@ from PySide6.QtCore import Qt
 
 from susapad.controller import exception
 from susapad.window import settings
-from susapad.window.main import component
+from susapad.window import __component
 from susapad import widget
 
 from . import ui
